refactoring for readability
